Core: Updated to 11.1.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/23624ed75abc8cbb77bb2e24a08e0836e4c41db5)

+++ Source/Game/BattleGrounds/BattleGround.cs
@@ -1216,8 +1216,8 @@ namespace Game.BattleGrounds
                     playerData.IsInWorld = true;
                     playerData.PrimaryTalentTree = (int)player.GetPrimarySpecialization();
                     playerData.Sex = (sbyte)player.GetGender();
-                    playerData.PlayerRace = player.GetRace();
-                    playerData.PlayerClass = (int)player.GetClass();
+                    playerData.PlayerRace = (sbyte)player.GetRace();
+                    playerData.PlayerClass = (sbyte)player.GetClass();
                     playerData.HonorLevel = (int)player.GetHonorLevel();
 
                     pvpLogData.Statistics.Add(playerData);
