+++ Source/Game/DataStorage/Structs/L_Records.cs
@@ -79,7 +79,7 @@ namespace Game.DataStorage
         public uint Id;
         public string Name;
         public string[] Texture = new string[6];
-        public ushort Flags;
+        public int Flags;
         public byte SoundBank;                                                // used to be "type", maybe needs fixing (works well for now)
         public uint SoundID;
         public uint SpellID;
