+++ Source/Game/Entities/Creature/CreatureData.cs
@@ -167,9 +167,9 @@ namespace Game.Entities
             stats.Flags[0] = (uint)creatureDifficulty.TypeFlags;
             stats.Flags[1] = creatureDifficulty.TypeFlags2;
 
-            stats.CreatureType = (int)CreatureType;
+            stats.CreatureType = (sbyte)CreatureType;
             stats.CreatureFamily = (int)Family;
-            stats.Classification = (int)Classification;
+            stats.Classification = (sbyte)Classification;
 
             for (uint i = 0; i < SharedConst.MaxCreatureKillCredit; ++i)
                 stats.ProxyCreatureID[i] = KillCredit[i];
